+++ src/SFA.DAS.ProviderRelationships.Web/ViewModels/AccountProviderLegalEntities/UpdateAccountProviderLegalEntityViewModel.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
 using SFA.DAS.Authorization;
 using SFA.DAS.ProviderRelationships.Application.Queries.GetAccountProviderLegalEntity.Dtos;
-using SFA.DAS.ProviderRelationships.DtosShared;
+using SFA.DAS.ProviderRelationships.Dtos;
 
 namespace SFA.DAS.ProviderRelationships.Web.ViewModels.AccountProviderLegalEntities
 {
