+++ src/SFA.DAS.ProviderRelationships.Web/ViewModels/AccountProviders/AddedAccountProviderViewModel.cs
@@ -1,5 +1,5 @@
 using System.ComponentModel.DataAnnotations;
-using SFA.DAS.ProviderRelationships.DtosShared;
+using SFA.DAS.ProviderRelationships.Dtos;
 
 namespace SFA.DAS.ProviderRelationships.Web.ViewModels.AccountProviders
 {
